+++ scripts/clean_up.sh
@@ -1,8 +1,10 @@
 find . -name "*.pyc" -exec rm -rf {} \;
 find . -name "__pycache__" -exec rm -rf {} \;
 find . -name "*.orig" -exec rm -rf {} \;
-rm -rf build/
-rm -rf dist/
 rm -rf .cache/
+rm -rf build/
+rm -rf builddocs/
+rm -rf dist/
+rm -rf deb_dist/
 rm src/tld.egg-info -rf
 rm builddocs.zip