+++ src/tld/__init__.py
@@ -6,10 +6,9 @@ from .utils import (
 )
 
 __title__ = 'tld'
-__version__ = '0.7.9'
-__build__ = 0x000013
+__version__ = '0.7.10'
 __author__ = 'Artur Barseghyan'
-__copyright__ = '2013-2017 Artur Barseghyan'
+__copyright__ = '2013-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'get_tld',