+++ src/tld/helpers.py
@@ -4,7 +4,7 @@ from .conf import get_setting
 
 __title__ = 'tld.helpers'
 __author__ = 'Artur Barseghyan'
-__copyright__ = '2013-2017 Artur Barseghyan'
+__copyright__ = '2013-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'project_dir',
@@ -14,9 +14,9 @@ __all__ = (
 
 def project_dir(base):
     """Project dir."""
-    TLD_NAMES_LOCAL_PATH_PARENT = get_setting('NAMES_LOCAL_PATH_PARENT')
+    tld_names_local_path_parent = get_setting('NAMES_LOCAL_PATH_PARENT')
     return os.path.abspath(
-        os.path.join(TLD_NAMES_LOCAL_PATH_PARENT, base).replace('\\', '/')
+        os.path.join(tld_names_local_path_parent, base).replace('\\', '/')
     )